refactor: created some packages to keep the same structure between tests and classes
