+++ owncloudData/src/main/java/com/owncloud/android/data/folderbackup/repository/OCFolderBackupRepository.kt
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.owncloud.android.data.folderbackup
+package com.owncloud.android.data.folderbackup.repository
 
 import com.owncloud.android.data.folderbackup.datasources.LocalFolderBackupDataSource
 import com.owncloud.android.domain.automaticuploads.FolderBackupRepository
